+++ src/pages/generate/components/header/FunnelHeader.css.ts
